Factor out shared chatbot utilities
