+++ src/rag_chatbot/user_manual/answer.py
@@ -35,6 +35,15 @@ def render_context(keep: List[Chunk]) -> str:
     )
 
 
+def filter_used_chunks(answer: str, chunks: List[Chunk]) -> List[Chunk]:
+    """Return only the chunks whose citations appear in the answer."""
+    used: List[Chunk] = []
+    for ch in chunks:
+        if ch.citation() in answer:
+            used.append(ch)
+    return used
+
+
 def answer_query(ix: Index, query: str) -> Tuple[str, List[Chunk]]:
     variants = multi_query_expand(query, ix.cfg)
 
@@ -51,5 +60,6 @@ def answer_query(ix: Index, query: str) -> Tuple[str, List[Chunk]]:
     sys_prompt = registry["answer_system"]
     user_prompt = registry["answer_user"].format(query=query, ctx=ctx)
     full_prompt = f"<|system|>\n{sys_prompt}\n<|user|>\n{user_prompt}"
-    ans = llm.invoke(full_prompt).content
-    return ans.strip(), kept
+    ans = llm.invoke(full_prompt).content.strip()
+    used = filter_used_chunks(ans, kept)
+    return ans, used
